Refactor util stuff
